% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soy_gff_from_soybase.R
\name{get_soy_gff_from_soybase}
\alias{get_soy_gff_from_soybase}
\title{Download the SoyBase Williams 82 GFF3 Annotation}
\usage{
get_soy_gff_from_soybase(dest_dir = "data", overwrite = FALSE)
}
\arguments{
\item{dest_dir}{Character string specifying the directory in which the
GFF3 annotation file will be downloaded and stored. The directory is
created recursively if it does not already exist. Default is
\code{"data"}.}

\item{overwrite}{Logical indicating whether an existing uncompressed GFF3
file should be replaced. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} containing the parsed Williams 82 soybean
  genome annotation.
}
\description{
Downloads the Williams 82 soybean genome annotation from SoyBase,
decompresses the GFF3 archive, and converts the annotation to a standard
data frame using \code{\link{get_gff}}.
}
\details{
This function is specifically designed for soybean (\emph{Glycine max})
and retrieves the Williams 82 genome annotation from SoyBase.

If the uncompressed GFF3 file already exists in \code{dest_dir} and
\code{overwrite = FALSE}, the existing file is used and no download is
performed.


The function downloads
\code{glyma.Wm82.gnm1.ann1.DvBy.gene_models_main.gff3.gz}
from SoyBase, decompresses the archive, removes the compressed file,
and passes the resulting GFF3 file to \code{\link{get_gff}} for parsing.
}
\examples{
\dontrun{
library(LDdecay)

gff_table <- get_soy_gff_from_soybase(
  dest_dir = "data"
)

head(gff_table)
}

}
\references{
SoyBase. Soybean Genome Database.
\url{https://www.soybase.org/}
}
