% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map}
\alias{map}
\title{Example Physical Marker Map}
\format{
A data frame with three columns:
\describe{
  \item{SNP}{SNP marker identifier.}
  \item{Chromosome}{Chromosome identifier.}
  \item{Position}{Physical position in base pairs.}
}
}
\usage{
map
}
\description{
A small soybean physical marker map corresponding to the SNP markers
included in the \code{geno} example dataset.
}
\examples{
data(map)
head(map)
}
\keyword{datasets}
