% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soy_gff_from_soybase.R
\name{get_soy_gff_from_soybase}
\alias{get_soy_gff_from_soybase}
\title{Download, Decompress, and Import SoyBase GFF3 Annotation Files}
\usage{
get_soy_gff_from_soybase(dest_dir = "data", overwrite = FALSE)
}
\arguments{
\item{dest_dir}{Character string pointing to the local directory where the downloaded 
GFF3 file should be stored. Default is \code{"data"}.}

\item{overwrite}{Logical. If \code{TRUE}, overwrites the file if it already exists 
locally in the target folder. Default is \code{FALSE}.}
}
\value{
A standard \code{data.frame} containing parsed genomic features (such as chromosomes, 
  start/end positions, strands, and associated metadata attributes). This function return gff3 file only for soybean. Download it manually for your crop of interest
}
\description{
Automatically connects to the authoritative SoyBase repository data collection, 
downloads the compressed Williams 82 genome annotation layer (.gff3.gz), and 
extracts it into a memory-optimized data frame ready for variant structural mapping.
}
\examples{
\dontrun{
library(LDdecay)

# Downloads from SoyBase, decompresses, and builds your matrix cleanly
gff_table <- get_soy_gff_from_soybase(dest_dir = "data")

# Preview the results instantly!
head(gff_table)
}

}
