% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_to_gene_mapping_using_gff3_annot.R
\name{snp_to_gene_mapping_using_gff3_annot}
\alias{snp_to_gene_mapping_using_gff3_annot}
\title{Map SNPs to Nearby Genomic Features using GFF3 Annotations}
\usage{
snp_to_gene_mapping_using_gff3_annot(
  snp_list,
  gff_path,
  map_data,
  window_bp = 50000
)
}
\arguments{
\item{snp_list}{A character vector of target SNP identifiers to query. e.g. list of all significant SNPs detected in GWAS}

\item{gff_path}{A character string pointing to the local paths of a standard \code{.gff3} file. Downloaded from https://data.soybase.org/Glycine/max/annotations/Wm82.gnm1.ann1.DvBy/}

\item{map_data}{A data frame containing physical map coordinates. Must include columns such as
\code{SNP}, \code{Chromosome}, and \code{Position}.}

\item{window_bp}{Numeric. The search radius in base-pairs flanking the variant. Default is \code{50000} (50kb).}
}
\value{
A polished \code{tibble} containing mapped variants, distance to midpoints, orientation, and descriptions.
}
\description{
Scans a physical window upstream and downstream of target SNPs to find overlapping
gene models extracted from a standard GFF3 file using high-performance non-equi joins.
}
\examples{
#' \dontrun{
library(LDdecay)
library(tidyverse)
# Define target test SNPs
test_snps <- c("ss715620779", "ss715620778", "ss715620770")
# Run mapping pipeline
test_mapping <- snp_to_gene_mapping_using_gff3_annot(
snp_list = test_snps,
gff_path = path/to/gff3/file, 
map_data = map, 
window_bp = 50000)
# View performance tables
print(test_mapping)
}

}
